+++ src/command/get_database.rs
@@ -1,10 +1,9 @@
 use hyper;
-use serde_json;
-use std;
 
 use client::{self, ClientState};
 use database::Database;
 use dbpath::DatabasePath;
+use dbtype;
 use error::{self, Error};
 use transport::{self, Command, Request};
 
@@ -57,41 +56,8 @@ impl<'a> Command for GetDatabase<'a> {
         match resp.status {
             hyper::status::StatusCode::Ok => {
                 let s = try!(client::read_json_response(&mut resp));
-                let mut resp_body = try!(client::decode_json::<serde_json::Value>(&s));
-                (|| {
-                    let dot = match resp_body.as_object_mut() {
-                        None => { return None; },
-                        Some(x) => x,
-                    };
-                    let doc_count = match dot.get("doc_count") {
-                        None => { return None; },
-                        Some(x) => match x.as_u64() {
-                            None => { return None; },
-                            Some(x) => x,
-                        },
-                    };
-                    let doc_del_count = match dot.get("doc_del_count") {
-                        None => { return None; },
-                        Some(x) => match x.as_u64() {
-                            None => { return None; },
-                            Some(x) => x,
-                        },
-                    };
-                    let db_name = match dot.get_mut("db_name") {
-                        None => { return None; },
-                        Some(x) => match *x {
-                            serde_json::Value::String(ref mut x) =>
-                                std::mem::replace(x, String::new()),
-                            _ => { return None; },
-                        },
-                    };
-                    Some(Database {
-                        doc_count: doc_count,
-                        doc_del_count: doc_del_count,
-                        db_name: db_name,
-                    })
-                })()
-                .ok_or(Error::UnexpectedContent { got: s } )
+                let db = try!(client::decode_json::<dbtype::Database>(&s));
+                Database::from_db_database(db)
             },
             hyper::status::StatusCode::NotFound =>
                 Err(error::new_because_not_found(&mut resp)),