+++ src/dbtype/mod.rs
@@ -2,8 +2,10 @@
 // as what the CouchDB API uses. Sometimes these layouts differ from what our
 // crate exports.
 
+mod database;
 mod viewresult;
 mod viewrow;
 
+pub use self::database::Database;
 pub use self::viewresult::ViewResult;
 pub use self::viewrow::ViewRow;